Add test_dropt, replace DROPT_NO_HELP with DROPT_NO_STRING_BUFFERS
Add a basic implementation of test_dropt.  Lots of testing code remains to
be added.

Also, DROPT_NO_HELP was ill-conceived since the error-reporting code used
code from dropt_string.c as well.  Refactor the code so that the error
message is generated lazily when trying to retrieve it instead of when the
error value is set.

And, as usual, fix some comments.  Also get rid of _CRT_NO_DEPRECATE by
using memcpy instead of strncpy.