Undo the previous change to pass the option name to dropt handlers
I think there's a better way to deal with custom errors.  Passing the
actual option names to handlers seems like overkill, making the handler
signature more cluttered unnecessarily for most clients, and it might be
susceptible to abuse.  I'd rather avoid it if I can.

The handlers won't need the dropt_context_t argument either, but it's
probably better to keep it for future flexibility.