Make dropt_parse not mutate argv
dropt_parse previously mutated argv (albeit non-destructively) temporarily
to tokenize strings like "--long=val" into separate "--long" and "val"
strings.

The straightforward way to avoid this would be to duplicate the original
string and to mutate the copy, but I've been trying to minimize memory
allocations.

Instead pass around a pointer and a string length to avoid needing to
insert a NUL-terminator.  Luckily the number of sites that needed to be
changed was fairly small, so hopefully this doesn't increase the code
complexity too much.