Make dropt emit an error for NULL arguments that aren't marked optional
If a dropt option expects a non-optional argument but uses an option
handler that accepts NULL, dropt was allowing that option to be used with
no arguments anyway.  Fix it.

Also adjust test_dropt to test for that case, and also replace its
"verbose" flag with a "quiet" flag.  The default behavior now is to print
diagnostics if command-line arguments are supplied.