Change the dropt directory structure
Move header files into a separate "include" directory and most of the source
files into a "src" directory. (I made an exception for the example files which
currently remain in the main dropt directory.)

Also, I had forgotten to add most of the new files from r469 to the packaging
script.