Remove package.cmd and make.cmd
I didn't really mean to commit package.cmd and make.cmd
(nor to include them as part of the dropt distribution).

package.cmd is a vestige from when I packaged the dropt
source manually; now that dropt is on github, that happens
automatically.

make.cmd was meant to be an internal convenience script for
me for building on Windows.